requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
cycler==0.11.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.5.3
multidict==6.0.4
numpy==1.21.6
openai==0.26.5
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
plotly==5.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tenacity==8.2.1
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
zipp==3.14.0
